Split GridItem into multiple files
