+++ frontend/src/views/admin/matching/applicant-view/grid/grid-item/applicant-star.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import classNames from "classnames";
-import { MatchableAssignment } from "../../types";
+import { MatchableAssignment } from "../../../types";
 import { BsStarFill } from "react-icons/bs";
-import { upsertMatch } from "../../actions";
-import { useThunkDispatch } from "../../../../../libs/thunk-dispatch";
+import { upsertMatch } from "../../../actions";
+import { useThunkDispatch } from "../../../../../../libs/thunk-dispatch";
 
 /**
  * A button for toggling applicant's "starred" status for the currently-selected position.
